Assignment_1/Classification/requirements.txt
pandas
numpy
scikit-learn
cupy-cuda11x
numba
ucimlrepo
psutil